new try

+++ ft_printf/includes/libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: sgertrud <msnazarow@gmail.com>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/03 02:21:10 by sgertrud          #+#    #+#             */
-/*   Updated: 2020/06/12 00:35:21 by sgertrud         ###   ########.fr       */
+/*   Updated: 2020/06/12 22:52:44 by sgertrud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,7 @@
 # include "libftprintf_utils.h"
 # include "length.h"
 # include "f_print.h"
+# include <stdbool.h>
 
 char			*parsing(char *s, t_format *format, va_list arg);
 int				ft_printf(const char *s, ...);
@@ -24,7 +25,7 @@ int				printing(t_format *format, va_list arg);
 void			clear(t_format *format);
 int				print_width(t_format *format, va_list arg);
 int				nbr(long long nb, int fd, t_format *format);
-int				nbr_f(double nb, int fd);
+int				nbr_f(long double nb, int fd);
 int				nbr_ull(unsigned long long nb, int fd, t_format *f);
 int				lbase(long long a, t_format *f, int base);
 int				ft_putstr_fd_f(char *s, int fd, t_format *format);
@@ -35,7 +36,7 @@ int				print_zeros(t_format *f);
 int				putstr(t_format *f, va_list arg);
 int				ft_putstr_w(wchar_t *s, int fd, t_format *f);
 int				is_float(char c);
-int				f_print(t_format *f, double d);
-int				e_print(t_format *f, double d);
+int				f_print(t_format *f,long double d);
+int				e_print(t_format *f,long double d);
 
 #endif
